add database.user tests

+++ app/database/_utils.py
@@ -18,49 +18,44 @@ def session_context():
         raise SQLAlchemyError(e)
 
 
-class Object:
+def create(model, *args, **kwargs):
     """
-    Base base class for manipulating database objects with SQLAlchemy.
+    Creates and returns a database object.
     """
+    obj = model(*args, **kwargs)
+    with session_context():
+        db.session.add(obj)
+    return obj
 
-    @staticmethod
-    def _create(model, *args, **kwargs):
-        """
-        Creates and returns a database object.
-        """
-        obj = model(*args, **kwargs)
+
+def get(model, key):
+    """
+    Retrieves and returns a database object for the specified model with the supplied identifier.
+    """
+    return model.query.get(key)
+
+
+def update(model, key, **kwargs):
+    """ 
+    Updates a database object.
+    The keyword arguments provided MUST correspond to column names for fields that will be updated.
+    
+    :type key: dict
+    """
+    if any(kwargs):
         with session_context():
-            db.session.add(obj)
-        return obj
+            model.query.filter_by(
+                **key
+            ).update(
+                {col: val for col, val in kwargs.items() if val is not None}
+            )
 
-    @classmethod
-    def _get(cls, model, key):
-        """
-        Retrieves and returns a database object for the specified model with the supplied identifier.
-        """
-        return model.query.get(key)
 
-    @staticmethod
-    def _update(model, key, **kwargs):
-        """ 
-        Updates a database object.
-        The keyword arguments provided MUST correspond to column names for fields that will be updated.
-        
-        :type key: dict
-        """
-        if any(kwargs):
-            with session_context():
-                model.query.filter_by(
-                    **key
-                ).update(
-                    {col: val for col, val in kwargs.items() if val is not None}
-                )
-
-    @classmethod
-    def _delete(cls, model, key):
-        """
-        Deletes a database record.
-        """
-        obj = cls._get(model, key)
+def delete(model, key):
+    """
+    Deletes a database record.
+    """
+    obj = get(model, key)
+    if obj is not None:
         with session_context:
             db.session.delete(obj)
